(spaceleader): Support omitting binding in leader-set-keys
Now if you supply a plist to leader-set-keys you can avoid supplying a
command to bind to, in which case the plist will be acted on but the
binding won't be interpreted. This is convenient for the use-package
declaration where I had unintentionally been unbinding leader-keys by
setting the command to nil.

For example `(leader-set-keys "ap" '(nil :wk "process"))` erases the
command at binding "ap" and then makes which key treat that binding as
"ap". Whereas `(leader-set-keys "ap" '(:wk "process"))` makes which key
substitute "ap" as "process" but leaves whatever binding was there unaffected.